Implement a gameboard factory module
The module stores the gameboard in a 3x3 array, with empty spaces
represented by a hyphen ('-').

The assignSpace method checks if the target space is occupied and
if the passed marker is valid (an 'x' or an 'o'). If so, then the element in the array
is replaced and the function returns a 0. However, if the space is
already occupied or the marker is invalid, then the function returns
a 1 or a 2 respectively.

The getMarkerAt method returns the character stored in the specified
array element.

The getBoardString method returns the current board state in a string
format.